updata doc and updata code
