+++ mmcls/core/__init__.py
@@ -1,3 +1,4 @@
 from .evaluation import *  # noqa: F401, F403
+from .export import *  # noqa: F401, F403
 from .fp16 import *  # noqa: F401, F403
 from .utils import *  # noqa: F401, F403
