+++ mmcls/core/export/__init__.py
@@ -1,0 +1,3 @@
+from .test import ONNXRuntimeClassifier
+
+__all__ = ['ONNXRuntimeClassifier']
