+++ tools/pytorch2onnx.py
@@ -37,7 +37,7 @@ def _demo_mm_inputs(input_shape, num_classes):
 def pytorch2onnx(model,
                  input_shape,
                  opset_version=11,
-                 dynamic_shape=False,
+                 dynamic_export=False,
                  show=False,
                  output_file='tmp.onnx',
                  do_simplify=False,
@@ -70,7 +70,7 @@ def pytorch2onnx(model,
     register_extra_symbolics(opset_version)
 
     # support dynamic shape export
-    if dynamic_shape:
+    if dynamic_export:
         dynamic_axes = {
             'input': {
                 0: 'batch',
@@ -121,7 +121,7 @@ def pytorch2onnx(model,
             output_file,
             input_shapes=input_shape_dic,
             input_data=input_dic,
-            dynamic_input_shape=dynamic_shape)
+            dynamic_input_shape=dynamic_export)
     if verify:
         # check by onnx
         import onnx
@@ -129,7 +129,7 @@ def pytorch2onnx(model,
         onnx.checker.check_model(onnx_model)
 
         # test the dynamic model
-        if dynamic_shape:
+        if dynamic_export:
             dynamic_test_inputs = _demo_mm_inputs(
                 (input_shape[0], input_shape[1], input_shape[2] * 2,
                  input_shape[3] * 2), model.head.num_classes)
@@ -176,7 +176,7 @@ def parse_args():
         default=[224, 224],
         help='input image size')
     parser.add_argument(
-        '--dynamic-shape',
+        '--dynamic-export',
         action='store_true',
         help='Whether to export ONNX with dynamic input shape. \
             Defaults to False.')
@@ -212,7 +212,7 @@ if __name__ == '__main__':
         input_shape,
         opset_version=args.opset_version,
         show=args.show,
-        dynamic_shape=args.dynamic_shape,
+        dynamic_export=args.dynamic_export,
         output_file=args.output_file,
         do_simplify=args.simplify,
         verify=args.verify)
